+++ app/Game/Engine/QuestRunConfig.php
@@ -12,6 +12,7 @@ final readonly class QuestRunConfig
         public int $questId,
         public int $stopRage = 2500,
         public bool $levelUp = false,
+        public bool $smart = false,
     ) {}
 
     /**
@@ -24,6 +25,7 @@ final readonly class QuestRunConfig
             questId: (int) ($config['quest_id'] ?? 0),
             stopRage: (int) ($config['stop_rage'] ?? 2500),
             levelUp: (bool) ($config['level_up'] ?? false),
+            smart: (bool) ($config['smart'] ?? false),
         );
     }
 
@@ -37,6 +39,7 @@ final readonly class QuestRunConfig
             'quest_id' => $this->questId,
             'stop_rage' => $this->stopRage,
             'level_up' => $this->levelUp,
+            'smart' => $this->smart,
         ];
     }
 }